feat: move utils functions
